[src/get_ld.c] Fixed build failure with GMP_NUMB_BITS == 8 due to the use of uint64_t in src/get_ld.c, while <stdint.h> was not included: replaced it by "unsigned long long", which does not need a specific header (an exact 64-bit type is not needed, we just need at least a 64-bit width, which unsigned long long is guaranteed to have). Note: unsigned long long may not be available with a pre-C99 compiler, but this is not worse than uint64_t. This limitation is currently OK as GMP_NUMB_BITS == 8 support is just for testing.
[From SVN r13950 (trunk)]

+++ src/get_ld.c
@@ -151,7 +151,7 @@ mpfr_get_ld (mpfr_srcptr x, mpfr_rnd_t rnd_mode)
         }
 #elif GMP_NUMB_BITS == 8
       {
-        uint64_t mant = 0;
+        unsigned long long mant = 0;
         int i;
         for (i = 0; i < 8; i++)
           mant |= ((unsigned long) tmpmant[i] << (8*i));
